Create Jamais Vu
The concept of this video was inspired by Tame Impala "Feels Like We Only Go Backwards" music video, designed by artist Leif Podhajsky.
I partnered with Para Chan to design this video with Processing.

Our concept for this project is to depict some of the important events and tragedies that have happened throughout history. We were inspired by the concept of psychedelic art which is reflected by revolution, politics, and historic and social events. The eye that is located on the centre, represents us, the humans, who were and still are, the observers of these events.
As for the visual aesthetics, we wanted to explore the sharp, vibrant colors of psychedelic art. And by adding features like repetition and movement, we tried to convey the feeling of depth.

The interaction that we included in our work is by mouseX and mouseY which controls the loops that are in the background. You can move the mouse around on the canvas and the form of the circles change by the location of your mouse. These form changes go from smooth circles to sharp and angular forms. 
What is the purpose of this interaction? The concept we wanted to explore was to go through the significant historic events and tragedies; kind of like traveling through time to take a second look at these events which made a huge change in human history. As we are observing these images, we get to express how we are feeling about them by changing the form of these loops.